feat: cut over agent programs to typed Python DAGs
Preserve original rowset laws through 162 paired matrix obligations, seven view scenarios, and host continuation tests. Add bounded Monty execution, typed host calls and occurrence streaming, incremental Python teaching and shared corrections across MCP, HTTP, Node, REPL, and evaluation.

Review repairs include signed literals, union/value preservation, packaged worker resolution, cancellation, stable aggregates, and parallel tracing isolation. Validation: 2548 library tests, 79 language tests, 15 view tests, 21 Node tests, 8 feasibility tests; workspace Clippy with warnings denied.

+++ crates/plasm-agent-core/src/plasm_plan_run/materialize.rs
@@ -36,7 +36,7 @@ pub(crate) async fn materialize_synthetic_node(
             if let crate::plasm_plan::ComputeOp::Limit { count } = &compute.compute.op {
                 plasm_runtime::coverage_after_explicit_take(
                     source_coverage,
-                    *count as usize,
+                    *count,
                     full_entities.len(),
                 )
             } else {
@@ -744,6 +744,8 @@ fn for_each_execution_result(
     result
 }
 
+// Archive context and fanout evidence are passed together, as in adjacent helpers.
+#[allow(clippy::too_many_arguments)]
 pub(crate) async fn archive_materialize_for_each_fanout(
     st: &PlasmHostState,
     es: &ExecuteSession,
